Another major overhaul of the MIDI import logic.
Getting closer to what we need for import.
The public interface is now shared pointers.  This keeps the raw pointer
fixed which addresses the issue when copying events.  For the common
use-cases the events will have a lifetime coincident with the playlist
(except de-duplicated and events deleted due to time combining).  This
may not be the most ideal use of C++ and surely wouldn't fly in anything
critical/safety related, but for a casual-use app where performance
isn't needed (performance is only needed during playback, it can take as
long as it wants for import) it should be OK.