Mostly bug squashing.
Moved all shared pointers out of the player thread.  Also significantly
reduced the time that the player thread keeps a lock (only 3 data
structures).  The UI thread is now responsible for all of the
"slow-copy" and "slow access" operations (IE list copy, shared_ptr
stuff, and locks) which won't degrate the user experience in the way
that it could degrade playback timing.