Capture state JIC.
I'm about to blow up a significant amount of the work to completely
restructure playlist loading and saving.  Data will be much more tied to
the UI playlist control elements.  The shared playlist structure will be
completely eliminated and the playlist entry structure will be
significantly altered.  After much thought, this is the best way to
marry the UI experience with the playlist sequence without needing to
maintain 2 separate structures.