Completely overhauled the playlist logic.
Playlist is now attached to the GUI controls representing the songs.
The GUI thread is responsible for enqueuing all songs to the player
thread when needed (based on signals coming from the UI or player
thread).  Internally, playlist is stored as a double-linked-list based
by song ID and the actual controls are stored in an unordered map.  GUI
elements don't actually swap - only their data components do so the GUI
doesn't have to re-bind the control widgets in the control layout.