Opening this back up for next release:
Next release will be known either as "Reformation" or "Advent".
Already tried doing the Linux port and convert to CMake.  This was a
complete fiasco.  See notes and attic_code for more details.  For the
immediate future it will remain VS2019 and MSW exclusively.

It looks like I'll be living with midifile as a submodule much longer
that I would have liked.  Going to try and remidy the problems.